backend/requirements.txt
Flask
flask-cors
openai-whisper
requests